:recycle: [TypeCentre] Migrate
