--- config/doctrine/TypeCentre.orm.yml
@@ -1,31 +1,0 @@
-App\Entity\TypeCentre:
-    type: entity
-    table: "typecentre"
-    repositoryClass: App\Repository\TypeCentreRepository
-    id:
-        id:
-            type: integer
-            id: true
-            generator:
-                strategy: AUTO
-    fields:
-        nom:
-            type: string
-            length: 255
-        createdAt:
-            type: datetime
-            column: createdAt
-            gedmo:
-                timestampable:
-                    on: create
-        updatedAt:
-            type: datetime
-            column: updatedAt
-            gedmo:
-                timestampable:
-                    on: update
-    oneToMany:
-        centres:
-            targetEntity: App\Entity\Attributes\Centre
-            mappedBy: typeCentre
-    lifecycleCallbacks: {  }
