+++ src/DataFixtures/FixturesPurger.php
@@ -22,7 +22,7 @@ use App\Entity\Attributes\MembreCentre;
 use App\Entity\Attributes\Note;
 use App\Entity\Attributes\RefreshToken;
 use App\Entity\Attributes\SharedDocument;
-use App\Entity\TypeCentre;
+use App\Entity\Attributes\TypeCentre;
 use App\Entity\User;
 use Doctrine\Common\DataFixtures\Purger\ORMPurgerInterface;
 use Doctrine\ORM\EntityManagerInterface;
